Added support for void __libc_init_array(void)

+++ lib-c/src/malloc.c
@@ -8,7 +8,7 @@
  * Copyright (C) 2014-2016  R. Stange <rsta2@o2online.de>
  * https://github.com/rsta2/circle/blob/master/lib/alloc.cpp
  */
-/* Copyright (C) 2017-2018 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2017-2023 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -201,7 +201,7 @@ void *calloc(size_t n, size_t size) {
 		*dst8++ = (uint8_t) 0;
 	}
 
-	assert(((void *)dst8 - (void *)p) == (n * size));
+	assert((size_t)((void *)dst8 - (void *)p) == (n * size));
 
 	return (void *) p;
 }
